+++ src/main/java/com/bg/bassheadsbg/model/dto/AddSubwooferDTO.java
@@ -35,7 +35,7 @@ public class AddSubwooferDTO {
 
     @Positive
     @NotNull
-    private byte impedance;
+    private float impedance;
 
     @Positive
     @NotNull
@@ -148,11 +148,11 @@ public class AddSubwooferDTO {
         this.numberOfCoils = numberOfCoils;
     }
 
-    public byte getImpedance() {
+    public float getImpedance() {
         return impedance;
     }
 
-    public void setImpedance(byte impedance) {
+    public void setImpedance(float impedance) {
         this.impedance = impedance;
     }
